refactor(order,ordertable): order와 ordertable 연관관계 끊기 위해 수정
- 인터페이스 추가,사용하여 OrderTableValidator에서 Order 애그리거트에 의존하는 부분 제거
