+++ src/main/java/kitchenpos/order/domain/OrderStateValidator.java
@@ -1,0 +1,24 @@
+package kitchenpos.order.domain;
+
+import kitchenpos.constants.ErrorMessages;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableConditionValidator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OrderStateValidator implements OrderTableConditionValidator {
+
+    private final OrderRepository orderRepository;
+
+    public OrderStateValidator(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    @Override
+    public void checkNotCompletedOrderExist(OrderTable orderTable) {
+        if (orderRepository.findByOrderTableId(orderTable.getId()).stream()
+                .anyMatch(Order::isOrderStatusNotComplete)) {
+            throw new IllegalArgumentException(ErrorMessages.CANNOT_CHANGE_EMPTY_IF_NOT_COMPLETED_ORDER_EXIST);
+        }
+    }
+}
