+++ src/main/java/kitchenpos/ordertable/domain/OrderTableConditionValidator.java
@@ -1,0 +1,5 @@
+package kitchenpos.ordertable.domain;
+
+public interface OrderTableConditionValidator {
+    void checkNotCompletedOrderExist(OrderTable orderTable);
+}
